Rename authMethodNames => idMethodNames
